# backend/core/internal/input/http/gen/types.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package gen

// CreateHabitRequest defines model for CreateHabitRequest.
type CreateHabitRequest struct {
	// Description Description of the Habit
	Description string `json:"description"`

	// Name Name of the Habit
	Name string `json:"name"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Error Human-readable error message
	Error string `json:"error"`
}

// Habit defines model for Habit.
type Habit struct {
	// Color Color of the Habit
	Color *string `json:"color,omitempty"`

	// CreatedAt Unix timestamp (milliseconds) when Habit was created
	CreatedAt *int64 `json:"createdAt,omitempty"`

	// Description Description of the Habit
	Description *string `json:"description,omitempty"`

	// DueDate Unix timestamp (milliseconds) when Habit is due, so-called goal
	DueDate *int64 `json:"dueDate,omitempty"`

	// Id Unique identifier for Habit
	Id HabitID `json:"id"`

	// Name Name of the Habit
	Name string `json:"name"`

	// UpdatedAt Unix timestamp (milliseconds) when Habit was last updated
	UpdatedAt *int64 `json:"updatedAt,omitempty"`

	// UserId Unique identifier for User
	UserId UserID `json:"userId"`
}

// HabitID Unique identifier for Habit
type HabitID = string

// UserID Unique identifier for User
type UserID = string

// ListHabitsParams defines parameters for ListHabits.
type ListHabitsParams struct {
	// Ids Filter by specific IDs
	Ids *[]HabitID `form:"ids,omitempty" json:"ids,omitempty"`
}

// CreateHabitJSONRequestBody defines body for CreateHabit for application/json ContentType.
type CreateHabitJSONRequestBody = CreateHabitRequest